Add corrent boost?

+++ current/src/constants.hpp
@@ -11,7 +11,7 @@
 static const int MAX_PARTS = 100;
 static const int N_SIGMA = 3;
 static const double PI = TMath::Pi();
-static const float INV_SQRT_2PI = TMath::Sqrt(2 * TMath::Pi());
+static const double INV_SQRT_2PI = TMath::Sqrt(2 * TMath::Pi());
 static const double D2R = PI / 180.0;
 static const int POSITIVE = 1;
 static const int NEGATIVE = -1;
